requirements.txt
accelerate
biopython
bitsandbytes
datasets
deepspeed
gitpython
huggingface_hub
langchain_community
langchain_core
markdownify
mashumaro
mwparserfromhell
newsapi-python
peft
pandera
pyyaml
sentencepiece
tokenizers
transformers==4.41.1
torch==2.0.1
wandb
unionai==0.1.18
flytekitplugins-envd==1.12.0
flytekitplugins-kfpytorch==1.12.0
flytekitplugins-huggingface==1.12.0
flytekitplugins-deck-standard==1.12.0
